struggling to stay C90 compatible... mostly (no %z printf specifier)
git-svn-id: svn://scm.orgis.org/mpg123/trunk@443 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/mpg123.c
@@ -827,7 +827,7 @@ int main(int argc, char *argv[])
 			if (split_dir_file(fname ? fname : "standard input",
 				&dirname, &filename))
 				fprintf(stderr, "\nDirectory: %s", dirname);
-			fprintf(stderr, "\nPlaying MPEG stream %zu of %zu: %s ...\n", pl.pos, pl.fill, filename);
+			fprintf(stderr, "\nPlaying MPEG stream %lu of %lu: %s ...\n", (unsigned long)pl.pos, (unsigned long)pl.fill, filename);
 
 #if !defined(GENERIC)
 {
